moved around recipe class, still doesnt actually work
