+++ src/main/java/dev/symphony/melody/Melody.java
@@ -3,6 +3,7 @@ package dev.symphony.melody;
 import dev.symphony.melody.config.MelodyConfigCondition;
 import dev.symphony.melody.item.ModItemGroups;
 import dev.symphony.melody.item.ModItems;
+import dev.symphony.melody.item.ModRecipes;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditionType;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditions;
 import dev.symphony.melody.network.MapBookOpenPayload;
@@ -38,7 +39,7 @@ public class Melody implements ModInitializer {
 		ModItemGroups.registerItemGroups();
 		ModItems.registerModItems();
 
-		RecipeRegistry.register();
+		ModRecipes.register();
 
 		MapBookOpenPayload.register();
 		MapBookSyncPayload.register();
