+++ src/main/resources/data/melody/recipes/map_book_addition.json
@@ -1,4 +1,4 @@
 {
-    "type": "fixedminecraft:map_book_addition",
+    "type": "melody:map_book_addition",
     "category": "misc"
 }
