+++ src/main/resources/data/melody/recipes/map_book_cloning.json
@@ -1,4 +1,4 @@
 {
-    "type": "fixedminecraft:map_book_cloning",
+    "type": "melody:map_book_cloning",
     "category": "misc"
 }
